birthday popup | mobile page connected with backend
This commit integrates the Birthday Popup and Birthday Mobile Page with the backend, enhancing the user experiences. The getBirthdayData function was implemented in HomePageModel.js to fetch birthday data, and HomePageController.js was updated to handle and display this data in a popup. Profile images are ensured to remain circular using CSS, and birthday messages were made more professional and context-aware. BirthdaysPageMobile.js was created to manage birthday notifications on the mobile page, with necessary updates to BirthdayPageMobile.html for scripts and styles.